23-4-2021 Date label, clock rendering pipeline
Wanted to implement the basics for the clock system, so
I had to re-do a bit of the data pipeline. The original plan was
to maybe have a possibility to choose between analog and digital clock.
It is still entirely possible - the current system just supports digital
clock.
Also added a label for date, so now there's a date and underneath it a
clock which gets updated every second. It's not super efficient, but it
works.